requirements.txt
python-dotenv==1.0.0
psycopg2==2.9.9
pandas==2.1.2
openpyxl==3.1.2
numpy
nbformat
matplotlib